Add bitcoinjs-browserify.
This should ideally be incorporated into the npm build, however it is not
immediately obvious how to do a patched install of a package using npm.
In this case we want to download bitcoinjs, apply our patch and use the
result.

Currently we use an unpatched bitcoinjs from npm for testing and a manually
generated bitcoinjs-browserify from this directory checked in
as ./static/external/bitcoinjs.min.js. This is not ideal.